+++ .all-contributorsrc
@@ -1,0 +1,15 @@
+{
+  "projectName": "react-if",
+  "projectOwner": "romac",
+  "repoType": "github",
+  "repoHost": "https://github.com",
+  "files": [
+    "README.md"
+  ],
+  "imageSize": 100,
+  "commit": false,
+  "commitConvention": "angular",
+  "contributors": [
+  ],
+  "contributorsPerLine": 7
+}
