--- .babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    "@babel/plugin-transform-modules-umd"
-  ]
-}
