+++ .editorconfig
@@ -1,4 +1,3 @@
-# http://editorconfig.org
 root = true
 
 [*]
@@ -9,6 +8,17 @@ charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
 
-[*.md]
-trim_trailing_whitespace = false
+[*.{js,ts,mjs}]
+indent_size = 2
+indent_style = space
+block_comment_start = /*
+block_comment = *
+block_comment_end = */
 
+[*.{yml,yaml}]
+indent_size = 2
+indent_style = space
+
+[*.{md,rmd,mkd,mkdn,mdwn,mdown,markdown,litcoffee}]
+tab_width = 2
+trim_trailing_whitespace = false
