--- .eslintrc
@@ -1,37 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "rules": {
-    "quotes": [0, "single", "avoid-escape"],
-    "key-spacing": 0,
-    "no-multi-spaces": 0,
-    "no-mixed-requires": 0,
-    "no-unused-vars": 0,
-    "no-undef": 0,
-    "new-cap": 0,
-    "no-underscore-dangle": 0,
-    "comma-spacing": 0,
-    "react/prop-types": 1,
-    "react/wrap-multilines": 1,
-    "react/jsx-uses-react": 1,
-    "react/jsx-key": 1,
-    "react/jsx-handler-names": 1,
-    "react/jsx-uses-vars": 1,
-    "react/no-did-mount-set-state": 1,
-    "react/no-did-update-set-state": 1,
-    "react/no-is-mounted": 1,
-    "react/no-danger": 1,
-    "react/no-deprecated": 1
-  },
-  "ecmaFeatures": {
-    "jsx": true,
-    "modules": true
-  },
-  "plugins": [
-    "react"
-  ],
-  "env": {
-    "browser": true,
-    "node": true,
-    "es6": true
-  }
-}
