+++ .github/CONTRIBUTING.md
@@ -1,0 +1,40 @@
+# Contributing
+
+To contribute to this repository, feel free to create a new fork of the repository and
+submit a pull request. We highly suggest [ESLint] to be installed
+in your text editor or IDE of your choice to ensure builds from GitHub Actions do not fail.
+
+1. Fork, clone, and select the **main** branch.
+2. Create a new branch in your fork.
+3. Make your changes.
+4. Ensure your linting and tests pass by running `yarn test && yarn lint`
+5. Commit your changes, and push them.
+6. Submit a Pull Request [here]!
+
+## Commands
+
+TSDX scaffolds your new library inside `/src`, and also sets up a [Parcel-based](https://parceljs.org) playground for it inside `/example`.
+
+The recommended workflow is to run TSDX in one terminal:
+
+```bash
+npm start # or yarn start
+```
+
+This builds to `/dist` and runs the project in watch mode so any edits you save inside `src` causes a rebuild to `/dist`.
+
+Then run the example inside another:
+
+```bash
+cd example
+npm i # or yarn to install dependencies
+npm start # or yarn start
+```
+
+The default example imports and live reloads whatever is in `/dist`, so if you are seeing an out of date component, make sure TSDX is running in watch mode like we recommend above. **No symlinking required**, we use [Parcel's aliasing](https://parceljs.org/module_resolution.html#aliases).
+
+To do a one-off build, use `npm run build` or `yarn build`.
+
+To run tests, use `npm test` or `yarn test`.
+
+[here]: https://github.com/romac/react-if/pulls
