+++ .gitignore
@@ -1,1 +1,5 @@
-node_modules/
+*.log
+.DS_Store
+node_modules
+.cache
+dist
