--- .travis.yml
@@ -1,21 +1,0 @@
-
-sudo: false
-
-language: node_js
-
-cache:
-  yarn: true
-  directories:
-    - "node_modules"
-
-node_js:
-  - "node"
-  - "lts/*"
-
-install:
-  - npm install
-
-script:
-  - npm run build
-  - npm run test
-
