+++ README.md
@@ -1,11 +1,8 @@
-
 # React If
 
 [![npm badge](http://img.shields.io/npm/v/react-if.svg)](https://www.npmjs.com/package/react-if)
-[![Build Status](https://travis-ci.org/romac/react-if.svg?branch=master&style=flat)](https://travis-ci.org/romac/react-if)
 [![Issues](http://img.shields.io/github/issues/romac/react-if.svg?style=flat)](https://github.com/romac/react-if/issues)
-![Status](https://img.shields.io/badge/status-inactive-lightgray.svg?style=flat)
-![License](https://img.shields.io/badge/license-mit-brightgreen.svg?style=flat)
+![License](https://img.shields.io/github/license/romac/react-if)
 [![Contact](https://img.shields.io/badge/contact-@__romac-blue.svg?style=flat)](https://twitter.com/_romac)
 
 Render React components conditionally.
@@ -14,41 +11,44 @@ Render React components conditionally.
 
 Take a look at the following presentational component, which contains a commonly used pattern for conditional rendering:
 
-```javascript
+```typescript
 const Bar = ({ name, age, drinkingAge }) => (
-    <div>
-        <Header />
-        {
-            age >= drinkingAge
-                ? <span className="ok">Have a beer, {name}!</span>
-                : <span className="not-ok">Sorry, {name}, you are not old enough.</span>
-        }
-        <Footer />
-    </div>
-)
+  <div>
+    <Header />
+    {age >= drinkingAge ? (
+      <span className="ok">Have a beer, {name}!</span>
+    ) : (
+      <span className="not-ok">Sorry, {name}, you are not old enough.</span>
+    )}
+    <Footer />
+  </div>
+);
 ```
 
 With `React-If` you can rewrite this into a more readable, expressive format:
 
-```javascript
+```typescript
 const Bar = ({ name, age, drinkingAge }) => (
-    <div>
-        <Header />
-        <If condition={ age >= drinkingAge }>
-            <Then><span className="ok">Have a beer, {name}!</span></Then>
-            <Else><span className="not-ok">Sorry, {name}, you are not old enough.</span></Else>
-        </If>
-        <Footer />
-    </div>
-)
+  <div>
+    <Header />
+    <If condition={age >= drinkingAge}>
+      <Then>
+        <span className="ok">Have a beer, {name}!</span>
+      </Then>
+      <Else>
+        <span className="not-ok">Sorry, {name}, you are not old enough.</span>
+      </Else>
+    </If>
+    <Footer />
+  </div>
+);
 ```
 
 ## Delaying evaluation of children / condition
 
-It is important to note that, because JavaScript is an eagearly evaluated language, children of both the `Then` and `Else` component and condition will be evaluated regardless of the value of the condition. Should that be an issue for performance reasons, one can wrap said children / condition in a [arrow function](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions), to delay evaluation of the children / condition, as in the following example:
+It is important to note that, because JavaScript is an eagerly evaluated language, children of both the `Then` and `Else` component and condition will be evaluated regardless of the value of the condition. Should that be an issue for performance reasons, one can wrap said children / condition in a [arrow function](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions), to delay evaluation of the children / condition, as in the following example:
 
-
-```javascript
+```typescript
 const renderData = (data) => {
   val computed = /* expensive computation */
   return <span>Here is the result: {computed}</span>;
@@ -92,87 +92,63 @@ And `this.props.bears.length` will not be called in the 2nd example.
 
 `npm install react-if`
 
-### Bower:
-
-`bower install react-if`
-
-```javascript
+```typescript
 // ES2015
-import { If, Then, Else, When, Unless, Switch, Case, Default } from 'react-if'
+import { If, Then, Else, When, Unless, Switch, Case, Default } from 'react-if';
 
 // CommonJS:
-const { If, Then, Else, When, Unless, Switch, Case, Default } = require('react-if')
-
-// Global
-var If   = ReactIf.If
-var Then = If.Then
-var Else = If.Else
-var When = If.When
-var Unless = If.Unless
-var Switch = If.Switch
-var Case = If.Case
-var Default = If.Default
+const { If, Then, Else, When, Unless, Switch, Case, Default } = require('react-if');
 ```
 
 ## Examples
 
 ## Swich/Case/Default
 
-```javascript
-import React from 'react'
-import { Switch, Case, Default } from 'react-if'
+```typescript
+import React from 'react';
+import { Switch, Case, Default } from 'react-if';
 
-const myNumber = 3
+const myNumber = 3;
 
 const Example = () => (
-    <div>
-        <Switch>
-          <Case condition={ myNumber === 9 }>
-            This will be displayed if condition is matched
-          </Case>
-          <Case condition={ myNumber > 1 }>
-            This will be displayed if condition is matched
-          </Case>
-          <Default>
-            This will be displayed if no Case have matching condition
-          </Default>
-        </Switch>
-    </div>
-)
+  <div>
+    <Switch>
+      <Case condition={myNumber === 9}>This will be displayed if condition is matched</Case>
+      <Case condition={myNumber > 1}>This will be displayed if condition is matched</Case>
+      <Default>This will be displayed if no Case have matching condition</Default>
+    </Switch>
+  </div>
+);
 ```
 
 ## Shorthands: When and Unless
 
-```javascript
-import React from 'react'
-import { When, Unless } from 'react-if'
+```typescript
+import React from 'react';
+import { When, Unless } from 'react-if';
 
-const someCondition = false
+const someCondition = false;
 
 const Example = () => (
-    <div>
-        <When condition={ someCondition }>
-            This will only be displayed, if the condition is TRUE
-        </When>
-    </div>
-)
+  <div>
+    <When condition={someCondition}>This will only be displayed, if the condition is TRUE</When>
+  </div>
+);
 
 const AnotherExample = () => (
-    <div>
-        <Unless condition={ someCondition }>
-            This will only be displayed, if the condition is FALSE
-        </Unless>
-    </div>
-)
+  <div>
+    <Unless condition={someCondition}>This will only be displayed, if the condition is FALSE</Unless>
+  </div>
+);
 ```
 
 ## API
 
 ### &lt;If /&gt;
 
-| Property      | Type    |
-| ------------- | ------- |
-| `condition`   | Boolean |
+| Property    | Type    |
+| ----------- | ------- |
+| `condition` | Boolean |
 
 If `condition` evaluates to `true`, renders the `<Then />` block will be rendered, otherwise renders the `<Else />` block. Either block may be omitted.
 
@@ -192,9 +168,9 @@ A container for `<Case condition={...}/>` and `<Default />` blocks. It will rend
 
 ### &lt;Case /&gt;
 
-| Property      | Type    |
-| ------------- | ------- |
-| `condition`   | Boolean |
+| Property    | Type    |
+| ----------- | ------- |
+| `condition` | Boolean |
 
 If the `Case` is the first one to have its `condition` evaluates to `true` inside the parent `<Switch />` it will be the only rendered.
 
@@ -213,3 +189,17 @@ A shorthand for `<If condition={...}><Else>...</Else></If>`. The same rules appl
 ## License
 
 **React If** is released under the [MIT license](http://romac.mit-license.org).
+
+## Contributors ✨
+
+Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/docs/en/emoji-key)):
+
+<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+<!-- markdownlint-restore -->
+<!-- prettier-ignore-end -->
+
+<!-- ALL-CONTRIBUTORS-LIST:END -->
+
+This project follows the [all-contributors](https://github.com/all-contributors/all-contributors) specification. Contributions of any kind welcome!
