--- demo/README.md
@@ -1,8 +1,0 @@
-
-# Demo
-
-- `npm install --dev`
-- `npm run build`
-- open `index.html` in the browser of your choice
-
-
