+++ demo/index.html
@@ -1,47 +1,31 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
-    <title>React If demo</title>
-    <script src="https://fb.me/react-0.14.6.js"></script>
-    <script src="https://fb.me/react-dom-0.14.6.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.23/browser.min.js"></script>
-    <script src="../lib/ReactIf.js"></script>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>Playground</title>
+
+    <style>
+      html,
+      body {
+        width: 100vw;
+        height: 100vh;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        align-content: center;
+        justify-content: center;
+        font-family: BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+        font-size: large;
+        color: beige;
+        background-color: #23272a;
+      }
+    </style>
   </head>
+
   <body>
-    <div id="example"></div>
-    <script type="text/babel">
-      const { If, Then, Else } = ReactIf;
-
-      const Beer = React.createClass({
-
-        getDefaultProps() {
-          return {
-            drinkingAge: 16 // Yay, Switzerland!
-          };
-        },
-
-        render() {
-          return (
-            <div id="some-id">
-              <If condition={ this.props.age >= this.props.drinkingAge }>
-                <Then><span>Have a beer, {this.props.name}!</span></Then>
-                <Else>
-                  {() => {
-                    return <span>Sorry, {this.props.name}, you are not old enough.</span>
-                  }}
-                </Else>
-              </If>
-            </div>
-          );
-        }
-
-      });
-
-      ReactDOM.render(
-        <Beer age={23} name="Romac" />,
-        document.getElementById('example')
-      );
-    </script>
+    <div id="root"></div>
+    <script src="./index.tsx"></script>
   </body>
 </html>
-
