+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testMatch: ['<rootDir>/test/**/*.test.tsx'],
+  setupFilesAfterEnv: ['<rootDir>/test/jest.setup.ts']
+};
