+++ src/Augments.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+  export interface ProcessEnv {
+    NODE_ENV: 'development' | 'production' | 'test';
+  }
+}
