+++ src/errorHandler.ts
@@ -1,0 +1,5 @@
+export const handleError = (errorMessage: string) => {
+  if (process.env.NODE_ENV !== 'production') {
+    throw new Error(errorMessage);
+  }
+};
