+++ src/index.tsx
@@ -1,0 +1,9 @@
+export * from './Case';
+export * from './Default';
+export * from './Else';
+export * from './If';
+export * from './Switch';
+export * from './Then';
+export * from './types';
+export * from './Unless';
+export * from './When';
