+++ tsconfig.eslint.json
@@ -1,0 +1,8 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "checkJs": true,
+        "allowJs": true
+    },
+    "include": ["src", "test", "demo", "./.eslintrc.js", "./jest.config.js", "./tsdx.config.js"],
+}
